refactoring 'validation_codes': moves code to _2fa submodule and improving security
